multiprocess simple framework as separate zmp package.
